requirements.txt
alembic==0.9.3
argcomplete==3.5.0
beautifulsoup4==4.11.1
bleach==2.0.0
blinker==1.4
boto3==1.26.74
botocore==1.29.165
bs4==0.0.1
certifi==2022.12.7
cffi==1.17.1
cfn-flip==1.3.0
charset-normalizer==2.0.12
click==6.7
cryptography==43.0.3
dominate==2.3.1
durationpy==0.7
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==5.0.0
Flask-HTTPAuth==3.2.3
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.0.4
Flask-Moment==0.5.1
Flask-PageDown==0.2.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.2
Flask-SSLify==0.1.5
Flask-WTF==0.14.2
greenlet==3.1.1
hjson==3.1.0
html5lib==0.999999999
idna==3.4
importlib-metadata==5.2.0
itsdangerous==0.24
Jinja2==2.9.6
jmespath==1.0.1
kappa==0.6.0
lxml==4.7.1
Mako==1.0.7
Markdown==2.6.8
MarkupSafe==2.0.1
numpy==2.0.2
pandas==2.2.3
placebo==0.9.0
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-decouple==3.6
python-editor==1.0.3
python-slugify==8.0.4
pytz==2021.3
PyYAML==6.0.2
requests==2.27.1
s3transfer==0.6.2
simplejson==3.18.3
six==1.10.0
soupsieve==2.3.1
SQLAlchemy==1.3.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.66.5
troposphere==4.8.1
typing_extensions==4.4.0
tzdata==2024.1
urllib3==1.26.8
visitor==0.1.3
webencodings==0.5.1
Werkzeug==0.12.2
WTForms==2.2
zappa==0.59.0
zipp==3.11.0
